refactor: Consolidate build targets
