--- CertificateSigningRequest (tvOS)/CertificateSigningRequest__tvOS_.h
@@ -1,18 +1,0 @@
-//
-//  CertificateSigningRequest__tvOS_.h
-//  CertificateSigningRequest (tvOS)
-//
-//  Created by Corey Baker on 12/11/20.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CertificateSigningRequest__tvOS_.
-FOUNDATION_EXPORT double CertificateSigningRequest__tvOS_VersionNumber;
-
-//! Project version string for CertificateSigningRequest__tvOS_.
-FOUNDATION_EXPORT const unsigned char CertificateSigningRequest__tvOS_VersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <CertificateSigningRequest__tvOS_/PublicHeader.h>
-
-
