--- CertificateSigningRequest (watchOS)/CertificateSigningRequest__watchOS_.h
@@ -1,18 +1,0 @@
-//
-//  CertificateSigningRequest__watchOS_.h
-//  CertificateSigningRequest (watchOS)
-//
-//  Created by Corey Baker on 12/11/20.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CertificateSigningRequest__watchOS_.
-FOUNDATION_EXPORT double CertificateSigningRequest__watchOS_VersionNumber;
-
-//! Project version string for CertificateSigningRequest__watchOS_.
-FOUNDATION_EXPORT const unsigned char CertificateSigningRequest__watchOS_VersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <CertificateSigningRequest__watchOS_/PublicHeader.h>
-
-
