+++ TestApp/Assets.xcassets/AccentColor.colorset/Contents.json
