+++ TestApp/Assets.xcassets/Contents.json
