+++ TestApp/ContentView.swift
@@ -1,0 +1,26 @@
+//
+//  ContentView.swift
+//  TestApp
+//
+//  Created by Corey Baker on 5/13/23.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundColor(.accentColor)
+            Text("Hello, world!")
+        }
+        .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
