--- TestHost/ContentView.swift
@@ -1,21 +1,0 @@
-//
-//  ContentView.swift
-//  TestHost
-//
-//  Created by Corey Baker on 12/11/20.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello, world!")
-            .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
