--- TestHostTV/Assets.xcassets/App Icon & Top Shelf Image.brandassets/Contents.json
@@ -1,32 +1,0 @@
-{
-  "assets" : [
-    {
-      "filename" : "App Icon - App Store.imagestack",
-      "idiom" : "tv",
-      "role" : "primary-app-icon",
-      "size" : "1280x768"
-    },
-    {
-      "filename" : "App Icon.imagestack",
-      "idiom" : "tv",
-      "role" : "primary-app-icon",
-      "size" : "400x240"
-    },
-    {
-      "filename" : "Top Shelf Image Wide.imageset",
-      "idiom" : "tv",
-      "role" : "top-shelf-image-wide",
-      "size" : "2320x720"
-    },
-    {
-      "filename" : "Top Shelf Image.imageset",
-      "idiom" : "tv",
-      "role" : "top-shelf-image",
-      "size" : "1920x720"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
